Feat: Controller을 기능에 따라 ViewController와 ApiController로 분리
